pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "task-inference"
version = "0.1.0"
description = "Abstractions for AI tasks following the Open Inference Protocol v2, with a HuggingFace Transformers reference implementation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pydantic>=2.0",
    "Pillow>=10.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
transformers = [
    "transformers==4.42.1",
    "torch>=2.0",
    "accelerate>=0.20",
    "timm>=0.9",
]
audio = [
    "soundfile>=0.12",
    "librosa>=0.10",
]
all = ["task-inference[transformers,audio]"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.6",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/your-org/task-inference"
Issues = "https://github.com/your-org/task-inference/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/task_inference"]

[tool.ruff]
src = ["src"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
